refactor(components): grouped the remaining sdk components together and cleaned up references
