+++ docs/cloud/worker-health.mdx
@@ -27,8 +27,7 @@ keywords:
   - temporal worker resources
 ---
 import { LANGUAGE_TAB_GROUP, getLanguageLabel } from '@site/src/constants/languageTabs';
-import SdkTabs from '@site/src/components';
-import { ReleaseNoteHeader } from '@site/src/components';
+import { ReleaseNoteHeader, SdkTabs } from '@site/src/components';
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
