+++ docs/develop/plugins-guide.mdx
@@ -13,7 +13,7 @@ tags:
   - AI Agents
 ---
 
-import SdkTabs from '@site/src/components';
+import { SdkTabs } from '@site/src/components';
 
 # Plugins
