+++ docs/develop/worker-performance.mdx
@@ -9,9 +9,8 @@ tags:
   - Performance
 ---
 
-import { CaptionedImage, ReleaseNoteHeader } from '@site/src/components';
+import { CaptionedImage, ReleaseNoteHeader, SdkTabs } from '@site/src/components';
 import { LANGUAGE_TAB_GROUP, getLanguageLabel } from '@site/src/constants/languageTabs';
-import SdkTabs from '@site/src/components';
 
 This page documents metrics and configurations that drive the efficiency of your Worker fleet.
 It provides coverage of performance metric families, Worker configuration options, Task Queue information, backlog counts, Task rates, and how to evaluate Worker availability.
@@ -248,7 +247,7 @@ These options define the maximum count of pollers performing poll requests on Wo
 
 #### SDK Examples
 
-<SdkTabs>
+<SdkTabs hideUnsupportedLanguages>
 <SdkTabs.Go>
 [Go SDK docs](https://pkg.go.dev/go.temporal.io/sdk/worker#PollerBehaviorAutoscalingOptions)
 ```go
