+++ docs/production-deployment/worker-deployments/serverless-workers/aws-lambda.mdx
@@ -18,8 +18,7 @@ tags:
   - AWS Lambda
 ---
 
-import SdkTabs from '@site/src/components/elements/SdkTabs';
-import { ReleaseNoteHeader } from '@site/src/components';
+import { ReleaseNoteHeader, SdkTabs } from '@site/src/components';
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
