+++ docs/production-deployment/worker-deployments/worker-versioning.mdx
@@ -18,7 +18,7 @@ tags:
 ---
 
 import { LANGUAGE_TAB_GROUP, getLanguageLabel } from '@site/src/constants/languageTabs';
-import SdkTabs from '@site/src/components';
+import { SdkTabs } from '@site/src/components';
 
 Worker Versioning is a Temporal feature that allows you to confidently deploy new changes to the Workflows running on
 your Workers without breaking them. Temporal enables this by helping you manage different builds or versions, formally
@@ -149,7 +149,7 @@ three new parameters, with different names depending on the language:
 
 Follow the example for your SDK below:
 
-<SdkTabs>
+<SdkTabs hideUnsupportedLanguages>
 <SdkTabs.Go>
 ```go
 buildID:= mustGetEnv("MY_BUILD_ID")
