+++ src/components/elements/Sdk/SdkGuideLinks/sdk-guide-links.module.css
@@ -9,7 +9,6 @@
   display: flex;
   align-items: center;
   gap: 10px;
-  text-decoration: none;
 }
 
 .iconWrapper {
