+++ src/components/elements/Sdk/index.js
@@ -1,0 +1,3 @@
+export * from './SdkGuideLinks'
+export * from './SdkLogosAsBlocks'
+export { default as SdkTabs } from './SdkTabs'
