+++ src/components/elements/index.js
@@ -4,3 +4,4 @@ export * from './Tile'
 export * from './Images'
 export * from './RelatedRead'
 export * from './ReleaseNoteHeader'
+export * from './Sdk'
