+++ src/components/index.js
@@ -6,10 +6,4 @@ export * from './Quickstart'
 export * from './ToolTipTerm';
 
 export { default as HomePageHero } from './elements/HomePageHero';
-export { SdkLogosAsBlocks } from './elements/SdkLogosAsBlocks';
-export { SdkGuideLinks } from './elements/SdkGuideLinks';
 export { default as PhotoCarousel } from './elements/PhotoCarousel';
-export { default as SdkTabs } from './elements/SdkTabs';
-
-// Extra export
-export { default } from './elements/SdkTabs';
